Added a RadioFrequency transmitter for the Raspberry Pi. The user can now have the Raspberry Pi transmit 433Mhz signals (given that the user properly set up a breadboard with the right components). I also added an interface so that the user can transmit via the Webserver.
Added TCGlobals; module aimed at providing global-level functions and variables.

Renamed TCEnumerations (The Chiem) to allow more general enumerations.

Added device type checking for device-specific operations (in our case, we want the raspberry pi to import the correct Pi Libraries such as GPIO and picam)

Changed CVCam to be a class in itself; In order to instantiate, you must call the factory functions. This makes it so that it is still treated as a singleton object, but the variables wouldn't be reinitialized during each module import.